requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9.0
django-modeltranslation>=0.18.0
django-rosetta>=0.9.0
celery>=5.3.0
redis>=5.0.0
Pillow>=9.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-extensions>=3.2.0
python-decouple>=3.8
django-cors-headers>=4.3.0
whitenoise>=6.6.0
gunicorn>=21.0.0
